permition

+++ app/Http/Controllers/DepartmentController.php
@@ -16,43 +16,65 @@ class DepartmentController extends Controller {
 
     public function for_id(Request $request)
     {
-        return Department::where('id', $request->id)->first();
+        $user = Auth::user();
+        if($user->role_id == 1)
+            return Department::where('id', $request->id)->first();
+        else
+            return ['status' => 'no permittion'];
     }
 
     public function create(Request $request)
     {
-        $dep = Department::create([
-            'name' => $request->name,
-            'full_name' => $request->full_name,
-            'email' => $request->email,
-            'phone' => $request->phone,
-            'phone_city' => $request->phone_city,
-        ]);
+        $user = Auth::user();
+        if($user->role_id == 1)
+        {
+            $dep = Department::create([
+                'name' => $request->name,
+                'full_name' => $request->full_name,
+                'email' => $request->email,
+                'phone' => $request->phone,
+                'phone_city' => $request->phone_city,
+            ]);
 
-        return $dep;
+            return $dep;
+        }
+        else
+            return ['status' => 'no permittion'];
     }
 
     public function update(Request $request)
     {
-        $dep = Department::where('id', $request->id)->first();
+        $user = Auth::user();
+        if($user->role_id == 1)
+        {
+            $dep = Department::where('id', $request->id)->first();
 
-        $dep->name = $request->name;
-        $dep->full_name = $request->full_name;
-        $dep->email = $request->email;
-        $dep->phone = $request->phone;
-        $dep->phone_city = $request->phone_city;
+            $dep->name = $request->name;
+            $dep->full_name = $request->full_name;
+            $dep->email = $request->email;
+            $dep->phone = $request->phone;
+            $dep->phone_city = $request->phone_city;
 
-        $dep->save();
+            $dep->save();
 
-        return $dep;
+            return $dep;
+        }
+        else
+            return ['status' => 'no permittion'];
     }
 
     public function delete(Request $request)
     {
-        $dep = Department::where('id', $request->id)->first();
+        $user = Auth::user();
+        if($user->role_id == 1)
+        {
+            $dep = Department::where('id', $request->id)->first();
 
-        $dep->delete();
+            $dep->delete();
 
-        return ['status' => 'success'];
+            return ['status' => 'success'];
+        }
+        else
+            return ['status' => 'no permittion'];
     }
 }
