+++ app/Http/Controllers/RoleController.php
@@ -2,6 +2,7 @@
 
 use App\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class RoleController extends Controller {
 
@@ -11,40 +12,66 @@ class RoleController extends Controller {
 
     public function list()
     {
-        return Role::get();
+        // $user = Auth::user();
+        // if($user->role_id == 1)
+            return Role::get();
+        // else 
+        //     return ['status' => 'no permittion'];
     }
 
     public function for_id(Request $request)
     {
-        return Role::where('id', $request->id)->first();
+        $user = Auth::user();
+        if($user->role_id == 1) //admin
+            return Role::where('id', $request->id)->first();
+        else
+            return ['status' => 'no permittion'];
     }
 
     public function create(Request $request)
     {
-        $role = Role::create([
-            'name' => $request->name,
-        ]);
+        $user = Auth::user();
+        if($user->role_id == 1) //admin
+        {
+            $role = Role::create([
+                'name' => $request->name,
+            ]);
 
-        return $role;
+            return $role;
+        }
+        else
+            return ['status' => 'no permittion'];
     }
 
     public function update(Request $request)
     {
-        $role = Role::where('id', $request->id)->first();
+        $user = Auth::user();
+        if($user->role_id == 1) //admin
+        {
+            $role = Role::where('id', $request->id)->first();
 
-        $role->name = $request->name;
+            $role->name = $request->name;
 
-        $role->save();
+            $role->save();
 
-        return $role;
+            return $role;
+        }
+        else
+            return ['status' => 'no permittion'];
     }
 
     public function delete(Request $request)
     {
-        $role = Role::where('id', $request->id)->first();
+        $user = Auth::user();
+        if($user->role_id == 1) //admin
+        {
+            $role = Role::where('id', $request->id)->first();
 
-        $role->delete();
+            $role->delete();
 
-        return ['status' => 'success'];
+            return ['status' => 'success'];
+        }
+        else
+            return ['status' => 'no permittion'];
     }
 }
